Update CLU instructions to install wget
wget is not necessarily installed on macOS systems by default and will need to be installed for 02_sqlserver_configure.sh to run correctly. 

Incidentally, the script continued to execute even though wget was not found on my system. Perhaps there's an opportunity to generalize CONT_OR_EXIT to be used for any command.